update overall function, remove overall_generate_hrd_function_dirty_test_ground.rs : get_hdr_phrases_function.rs
